lesson 1-4 double to float done

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -4,7 +4,7 @@ fun main() {
     
     val length: Long = 40868600000
     val age: Byte = 27
-    val partOfDay: Double = 0.075
+    val partOfDay: Float = 0.075F
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
